requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
cryptography